% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNLP.R
\name{Create_Vocab_Document_Term_Matrix}
\alias{Create_Vocab_Document_Term_Matrix}
\title{Input a dataframe of corpus and output document term Matrix and vocab
  specify column with text to use}
\usage{
Create_Vocab_Document_Term_Matrix(input, col = 1,
  words_to_remove = NULL, return_embedding_params = F)
}
\arguments{
\item{input}{A dataframe of a text corpus}

\item{col}{The column of the dataframe you wish to use; defaults to one}

\item{words_to_remove}{The words you wish to not be included from Vocab_to_Remove function.
defaults to NULL}

\item{return_embedding_params}{Whether to to include the itokens and vocab_vectorizer in the output
which is needed to pass to train_embeddings if you wish to train embeddings on your own data.
Defaults to false.}
}
\value{
Returns a list of the document term matrix and vocabulary with term counts.
}
\description{
Input a dataframe of corpus and output document term Matrix and vocab
  specify column with text to use
}
\examples{
Create_Vocab_Document_Term_Matrix(MyData,col=1,words_to_remove=c('kale','lump'))
}
