% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNLP.R
\name{CreateEmbeddingMatrix}
\alias{CreateEmbeddingMatrix}
\title{Create a matrix of embeddings}
\usage{
CreateEmbeddingMatrix(DocTermMatrix, Embeddings, Vocabulary, LogInverse)
}
\arguments{
\item{DocTermMatrix}{The document term matrix}

\item{Embeddings}{The matrix of embeddings (e.g. Glove) where rows are words}

\item{Vocabulary}{The Vocab list or dataframe from Create_Vocab_DocumentTermMatrix}

\item{LogInverse}{whether to weight words using LogInverse; defaults to FALSE}
}
\value{
Returns the average embedding of each document for each embedding dimension.
}
\description{
Create a matrix of embeddings
}
\examples{
CreateEmbeddingMatrix(MyDocTermMatrix, Glove.200, Vocab, LogInverse=T)
}
