% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNLP.R
\name{find_best_neural_net}
\alias{find_best_neural_net}
\title{Does a search for the best neural net to fit the data using a restricted h2o.automl function.
See this function in h2o documentation for more info.
You will need the h2o package. And you may have to download an older version of java for it to work.}
\usage{
find_best_neural_net(Embeddings, dependent_var, max_minutes = 10,
  stopping_tol = 0.005, n_folds = 5)
}
\arguments{
\item{Embeddings}{The embedding dataframe or matrix}

\item{dependent_var}{The variable used to predict, must be a factor}

\item{max_minutes}{The maximum run time for model search if stopping tolerance not reached.
For larger data sets you may need multiple hours. Defaults to 10}

\item{stopping_tol}{The stopping tolerance for model search. Defaults to .005}

\item{nfolds}{The number of folds for k-fold cross-validation. Defaults to 5}
}
\value{
Returns the best model
}
\description{
Does a search for the best neural net to fit the data using a restricted h2o.automl function.
See this function in h2o documentation for more info.
You will need the h2o package. And you may have to download an older version of java for it to work.
}
\examples{

}
