% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNLP.R
\name{Remove_Bad_Chars}
\alias{Remove_Bad_Chars}
\title{Removes Numbers and Punctuation characters}
\usage{
Remove_Bad_Chars(string)
}
\arguments{
\item{string}{An input that may be a string or dataframe column}
}
\value{
returns the input in same format but without punctuation or numbers
}
\description{
This function removes punctuation and numbers
}
\examples{
Remove_Bad_Chars('1 plus 1.1 = 2.1')
}
