% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNLP.R
\name{remove_gendered_words}
\alias{remove_gendered_words}
\title{Remove common gender pronouns and words from a corpus}
\usage{
remove_gendered_words(x, remove_familial = T)
}
\arguments{
\item{x}{The input of strings(datframe,vector,etc)}

\item{remove_familial}{Removes familial gender terms: aunt, uncle, mom, dad,
mother, father, husband, wife, son, daughter, niece, nephew, grandmother,
grandfather. Defaults to true}
}
\value{
Returns the input but without common gendered words
}
\description{
Remove common gender pronouns and words from a corpus
}
\examples{
remove_gendered_words(MyCorpus)
}
