% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyNLP.R
\name{Vocab_to_Remove}
\alias{Vocab_to_Remove}
\title{Gets words of certain vocaublary from an input soruce (String, column, etc)
  which you may wish to remove later on.}
\usage{
Vocab_to_Remove(RemovableData)
}
\arguments{
\item{RemovableData}{An input that may be a string or dataframe column}
}
\value{
Returns all the individual words that you do not want
}
\description{
Gets words of certain vocaublary from an input soruce (String, column, etc)
  which you may wish to remove later on.
}
\examples{
Vocab_to_Remove(c('one direction','kale','russia'))
}
